requirements.txt
fastmcp==2.12.3
sendgrid
httpx
python-dotenv
pydantic==2.11.9
pydantic-settings==2.10.1